Time: 19 ms (81.59%) | Memory: 12.9 MB (59.30%) - LeetSync

+++ 843-binary-trees-with-factors/binary-trees-with-factors.cpp
@@ -10,8 +10,8 @@ public:
         }
         for(int i=0;i<n;i++){
             for(int j=0;j<i;j++){
-                if(arr[i]%arr[j]==0 && mapy.find((long long)arr[i]/arr[j])!=mapy.end()){
-                    mapy[arr[i]]+=(mapy[arr[j]]%MOD)*(mapy[(long)arr[i]/arr[j]]%MOD)%MOD;
+                if(arr[i]%arr[j]==0 && mapy.find(arr[i]/arr[j])!=mapy.end()){
+                    mapy[arr[i]]+=(mapy[arr[j]]%MOD)*(mapy[arr[i]/arr[j]]%MOD)%MOD;
                 }
             }
         }
